Implemented the minimum ORF length functionality
This feature, required by the assignment, allows the user to specify a minimum length for the ORFs (s)he would like to view.  A new function was added to Parser_revised.py to prompt the user for this minimum length, and that value was then passed to Printout_Function.py and used as a filter (we won't print any ORFs that are shorter than this value).